I created connection between 2 viewControllers
